initscripts: just try to mount rpc_pipefs always
Do not rely on modprobe.d hacks to mount the filesystem after loading the module.
Mount the filesystem directly. The kernel will ask for the module if needed.

This avoids race conditions between the initscripts running in parallel and it
works with built-in sunrpc too.

Later we can consider passing the responsibility to mount rpc_pipefs
to a systemd {auto,}mount unit pair.

https://bugzilla.redhat.com/show_bug.cgi?id=702707

Signed-off-by: J. Bruce Fields <bfields@redhat.com>