Rely on crypto module autoloading in init scripts
SECURE_NFS_MODS isn't necessary any more as we'll autoload any crypto
modules we need, so rpcsec_gss_krb5 is all we need to load.  Even that
we could autoload as well, but that will require a kernel fix, and to
support older kernels we should wait till that fix has been in for a
while.

Signed-off-by: J. Bruce Fields <bfields@redhat.com>