Clean up lockd configuration
Configuring lockd to listen on a particular set of ports has always been
an iffy proposition. We've tried for years to set it up so that you can
configure it via /etc/sysconfig/nfs, but there are some cases where it
may not get configured properly.

If lockd is started via an nfs mount prior to starting the
nfs-lock.service then it'll end up listening on a random, ephemeral port
instead of the one configured.

Instead of trying to do this, let's formally deprecate the LOCKD_TCPPORT
and LOCKD_UDPPORT parameters in /etc/sysconfig/nfs. The version of it
doesn't even have commented out versions of them anyway. Replace them
with a note that tells admins that they should set up static ports in
/etc/modprobe.d/lockd.conf.

Add an /etc/modprobe.d/lockd.conf with a commented out options directive
to make it easy for them to do so.

Signed-off-by: Jeff Layton <jlayton@primarydata.com>
Signed-off-by: Steve Dickson <steved@redhat.com>